requirements.txt
discord.py[voice]==1.2.5
discord-py-slash-command
pip
python-dotenv
